adjustments made to code lead to very similar results, but not always exact matches. Investigate if this is due to numerical instability or a bug

+++ include/qalgorithms_qpeaks.h
@@ -14,7 +14,7 @@
 namespace qAlgorithms
 {
     std::vector<RegCoeffs> findCoefficients(
-        const std::vector<float> intensity_log,
+        const std::vector<double> intensity_log,
         const size_t scale,      // maximum scale that will be checked. Should generally be limited by peakFrame
         const size_t numPeaks,   // only > 1 during componentisation (for now? @todo)
         const size_t peakFrame); // how many points are covered per peak? For single-peak data, this is the length of intensity_log
